test merge code

+++ src/components/auth/Register.js
@@ -5,8 +5,21 @@ import { MoviesContext } from "../../App";
 function Register() {
     const { accounts, setAccounts } = useContext(MoviesContext)
 
+    const [gender, setGender] = useState(true)
+
     const isBlank = (param) => {
         if (typeof param === 'undefined' || param.toString() === '')
+            return true
+        return false
+    }
+
+    const localAccounts = JSON.parse(localStorage.getItem('accounts'))
+
+
+    const checkExist = (userName) => {
+        const account = localAccounts.find(e => e.userName == userName)
+
+        if (account)
             return true
         return false
     }
@@ -30,28 +43,45 @@ function Register() {
             if (!isBlank(password)) {
                 setMessagePassword('')
                 if (!isBlank(fullName)) {
-                    setMessageFullName('')
-                    const newData = {
-                        id: accounts[accounts.length - 1].id + 1,
-                        userName: userName,
-                        password: password,
-                        fullName: fullName
-                    }
-                    const newAccounts = [...accounts, newData];
+                    if (!checkExist(userName)) {
+                        setMessageFullName('')
+                        const newData = {
+                            id: accounts[accounts.length - 1].id + 1,
+                            userName: userName,
+                            password: password,
+                            fullName: fullName,
+                            isPrimeum: false,
+                            active: true,
+                            gender: gender,
+                            role: 'user'
+                        }
+                        const newAccounts = [...accounts, newData];
 
-                    localStorage.setItem('accounts', JSON.stringify(newAccounts))
-                    setAccounts(newAccounts)
-                    setMessage('Đăng Ký Thành Công !')
+                        localStorage.setItem('accounts', JSON.stringify(newAccounts))
+                        setAccounts(newAccounts)
+                        setMessage('Đăng Ký Thành Công !')
+
+                        userNameField.current.value = ''
+                        passwordField.current.value = ''
+                        fullNameField.current.value = ''
+
+                    } else {
+                        setMessage('Tài Khoản Đã Tồn Tại !')
+                    }
+
                 } else {
                     setMessageFullName('FullName không được để trống !')
+                    fullNameField.current.focus()
 
                 }
             } else {
                 setMessagePassword('Password không được để trống !')
+                passwordField.current.focus()
 
             }
         } else {
             setMessageUserName('UserName không được để trống !')
+            userNameField.current.focus()
 
         }
 
@@ -72,23 +102,34 @@ function Register() {
                                         messageUserName !== '' ? <p className="text-danger">{messageUserName}</p> : ''
                                     }
                                 </div>
-                                <div className="form-group mt-2">
+                                <div className="form-group mt-3">
                                     <label forhtml="password">Password: </label>
                                     <input type="password" ref={passwordField} name="password" id="password" className="form-control mt-2" />
                                     {
                                         messagePassword !== '' ? <p className="text-danger">{messagePassword}</p> : ''
                                     }
                                 </div>
-                                <div className="form-group mt-2">
+                                <div className="form-group mt-3">
                                     <label forhtml="fullName">Full Name: </label>
                                     <input type="text" ref={fullNameField} name="fullName" id="fullName" className="form-control mt-2" />
                                     {
                                         messageFullName !== '' ? <p className="text-danger">{messageFullName}</p> : ''
                                     }
                                 </div>
+                                <div className="form-group mt-3 d-flex">
+                                    <label forhtml="fullName">Gender: </label>
+                                    <div className="form-check mx-4">
+                                        <input class="form-check-input" type="radio" name="gender" id="male" value="true" defaultChecked onClick={() => setGender(true)} />
+                                        <label class="form-check-label" forhtml="male">Male</label>
+                                    </div>
+                                    <div className="form-check">
+                                        <input class="form-check-input" type="radio" name="gender" id="female" value="" onClick={() => setGender(false)} />
+                                        <label class="form-check-label" forhtml="female">Female</label>
+                                    </div>
+                                </div>
                                 <div className="form-group mt-3">
                                     {
-                                        message !== '' ? <p className="text-success">{message}</p> : ''
+                                        message === 'Tài Khoản Đã Tồn Tại !' ? <p className="text-danger">{message}</p> : <p className="text-success">{message}</p>
                                     }
                                     <button type="button" className="btn btn-primary" onClick={handleRegister}>Đăng Ký</button>
                                 </div>
